% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMD.R
\docType{data}
\name{DMD}
\alias{DMD}
\title{A dataset on Duchenne Muscular Dystrophy (DMD)}
\format{A data frame with 194 observations of 5 variables.
\itemize{
\item CK: the measurements of biomarker CK,
\item H: the measurements of biomarker H,
\item LD: the measurements of biomarker LD,
\item PK: the measurements of biomarker PK,
\item Status: 0 =  non-carrier of DMD and 1 = carrier of DMD.
}}
\usage{
data(DMD)
}
\description{
This dataset contians complete data of four biomarkers, namely,
creatinekinase (CK), hemopexin (H), lactate dehydroginase (LD), pyruvate kinase (PK),
and status whether the object is a carrier of DMD.
See \insertCite{andrews2012data;textual}{YoudenDRM} for more details.
}
\references{
\insertRef{andrews2012data}{YoudenDRM}
}
