% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRM.R
\name{DRM}
\alias{DRM}
\title{Density Ratio Model (DRM)}
\usage{
DRM(x, y, qt, r = NULL, totalSize, method = "optimal")
}
\arguments{
\item{x}{observed sample from \eqn{F_0}.}

\item{y}{observed sample from \eqn{F_1}.}

\item{qt}{pre-sepecified basis functions of t in the exponential term. For example, \code{qt = c("t","log(t)")}.}

\item{r}{the value of the LLOD. The default case is where no LLOD exists.}

\item{totalSize}{the total sample sizes \eqn{(n_0,n_1)}. When there is no LLOD (\code{r = NULL}), this argument is optional. See `Details'.}

\item{method}{the method used to fit the DRM. The options include \code{"glm"}, \code{"multinom"} and \code{"optimal"}.
The default setting is \code{"optimal"}. See `Details'.}
}
\value{
The function returns a list containing the following components:
\itemize{
\item prob_obs: the estimates of \eqn{\zeta_0 = P(x_{01}\geq r)} and \eqn{\zeta_1 = P(x_{11}\geq r)}. If there is no LLOD, it returns (1,1).
\item sample_obs: the combined observed samples.
\item coef: the estimates of parameters \eqn{\alpha} and \eqn{\boldsymbol{\beta}}.
\item P0: the estimated density of \eqn{F_0} at each observed data point.
\item P1: the estimated density of \eqn{F_1} at each observed data point.
}
}
\description{
Fit a semiparametric two-sample DRM based on data with or without the lower limit of detetion (LLOD).
}
\details{
Denote \eqn{\{X_{01},\ldots,X_{0n_0}\}} and \eqn{\{X_{11},\ldots,X_{1n_1}\}} as two independent random samples coming from the healthy and diseased populations, respectively.
Let \eqn{F_0} and \eqn{F_1} denote the cumulative distribution functions of the healthy population and the diseased population.
\insertCite{yuan2021semiparametric}{YoudenDRM} porposed to link \eqn{F_0} and \eqn{F_1} by a DRM:
\deqn{dF_1(x) = \exp\{\alpha + \boldsymbol{\beta}^\top \boldsymbol{q}(x)\}dF_0(x),}
where \eqn{dF_i} denote the density of \eqn{F_i}; \eqn{\alpha} and \eqn{\boldsymbol{\beta}} are unknown parameters for the DRM;
\eqn{\boldsymbol{q}(x)} is a pre-specified, non-trivial function of dimension \eqn{d};
 the baseline distribution \eqn{F_0} is unspecified.

Suppose the biomarker has a LLOD, denoted as \eqn{r}. Let \eqn{m_0} and \eqn{m_1} be the numbers of observations above the LLOD \eqn{r} in the healthy and diseased groups, respectively.
Without loss of generality, we assume the first \eqn{m_i} observations in sample \eqn{i}, \eqn{X_{i1},\ldots,X_{im_i}} are above the LLOD.
According to \insertCite{yuan2021semiparametric;textual}{YoudenDRM},
the estimators of \eqn{\alpha} and \eqn{\boldsymbol{\beta}} maximize the following empirical log-likelihood function
\deqn{\ell_{n}(\alpha,\boldsymbol{\beta} ) = \sum_{j = 1}^{m_1} \{\alpha + \boldsymbol{\beta}^\top \boldsymbol{q}(X_{1j})\} - \sum_{i = 0}^1\sum_{j = 1}^{m_i}\log\left[1+\frac{n_1}{n_0} \exp\{\alpha + \boldsymbol{\beta}^\top \boldsymbol{q}(X_{ij})\}\right].}

Since the DRM is equivalent to the logistic regression model with some justification of \eqn{\alpha},
we can use function \code{glm} or \code{multinom} to solve for parameters \eqn{\alpha} and \eqn{\boldsymbol{\beta}} instead of maximizing the above likelihood function.
The method \code{optimal} estimates the parameters \eqn{\alpha} and \eqn{\boldsymbol{\beta}} by \code{glm} or \code{multinom}, whichever gives the larger likelihood \eqn{\ell_{n}(\alpha,\boldsymbol{\beta})}.
}
\references{
\insertRef{yuan2021semiparametric}{YoudenDRM}
}
