% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inference.R
\name{DRMci}
\alias{DRMci}
\title{Confidence intervals for the Youden index and optimal cutoff point}
\usage{
DRMci(x, y, qt, r = NULL, totalSize, method = "optimal", logit = T,
  level = 0.95)
}
\arguments{
\item{x}{observed sample from \eqn{F_0}.}

\item{y}{observed sample from \eqn{F_1}.}

\item{qt}{pre-sepecified basis functions of t in the exponential term. For example, \code{qt = c("t","log(t)")}.}

\item{r}{the value of the LLOD. The default case is where no LLOD exists.}

\item{totalSize}{the total sample sizes \eqn{(n_0,n_1)}. When there is no LLOD (\code{r = NULL}), this argument is optional. See \code{help(DRM)} `Details'.}

\item{method}{the method used to fit the DRM. The options include \code{"glm"}, \code{"multinom"} and \code{"optimal"}.
The default setting is \code{"optimal"}. See \code{help(DRM)} for details.}

\item{logit}{whether the logit transformation is applied when constructing confidence interval for the Youden index.}

\item{level}{confidence level of the confidence interval. The default value is 0.95.}
}
\value{
The function returns a list containing the following components:
\itemize{
\item  CI_J: confidence interval of the Youden index.
\item  CI_c :confidence interval of the optimal cutoff point.
}
}
\description{
Construct CIs for the Youden index and its corrreponding cutoff point under DRMs based on the provided data.
}
\examples{
#DMD is the dataset of data application in \\insertCite{yuan2021semiparametric;textual}{YoudenDRM}.
x = DMD$CK[DMD$Status == 0]
y = DMD$CK[DMD$Status == 1]
DRMci(x,y,qt = "t")
#Date generation
set.seed(123456)
x = rlnorm(50, meanlog = 2.5, sdlog = sqrt(0.09))
y = rlnorm(50, meanlog = 2.87,sdlog = sqrt(0.25))
#Create the LLOD
r = qlnorm(0.15,meanlog = 2.5, sdlog = sqrt(0.09))
#Observed samples
x = x[x>=r]
y = y[y>=r]
# The basis function
qt = c("log(t)","log(t)^2")
DRMci(x,y,qt,r,totalSize = c(50,50))
}
\references{
\insertRef{yuan2021semiparametric}{YoudenDRM}
}
