% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inference.R
\name{DRMest}
\alias{DRMest}
\title{Point estimation of the Youden index and optimal cutoff point}
\usage{
DRMest(x, y, qt, r = NULL, totalSize, method = "optimal")
}
\arguments{
\item{x}{observed sample from \eqn{F_0}.}

\item{y}{observed sample from \eqn{F_1}.}

\item{qt}{pre-sepecified basis functions of t in the exponential term. For example, \code{qt = c("t","log(t)")}.}

\item{r}{the value of the LLOD. The default case is where no LLOD exists.}

\item{totalSize}{the total sample sizes \eqn{(n_0,n_1)}. When there is no LLOD (\code{r = NULL}), this argument is optional. See \code{help(DRM)} `Details'.}

\item{method}{the method used to fit the DRM. The options include \code{"glm"}, \code{"multinom"} and \code{"optimal"}.
The default setting is \code{"optimal"}. See \code{help(DRM)} for details.}
}
\value{
The function returns a list containing the following components:
\itemize{
\item J:  estimate of the Youden index,
\item sdJ: standard devatation of the estimator of the Youden index,
\item c: sdcestimate of the optimal cutoff point,
\item sdc: standard devatation of the estimator of the optimal cutoff point.
}
}
\description{
Estimate the Youden index and its corrreponding cutoff point and standard deviation of estimators under DRM based on the provided data.
}
\examples{
#DMD is the dataset of data application in \\insertCite{yuan2021semiparametric;textual}{YoudenDRM}.
x = DMD$CK[DMD$Status == 0]
y = DMD$CK[DMD$Status == 1]
DRMest(x,y,qt = "t")
#Date generation
set.seed(123456)
x = rlnorm(50, meanlog = 2.5, sdlog = sqrt(0.09))
y = rlnorm(50, meanlog = 2.87,sdlog = sqrt(0.25))
#Create the LLOD
r = qlnorm(0.15,meanlog = 2.5, sdlog = sqrt(0.09))
#Observed samples
x = x[x>=r]
y = y[y>=r]
# The basis function
qt = c("log(t)","log(t)^2")
DRMest(x,y,qt,r,totalSize = c(50,50))
}
\references{
\insertRef{yuan2021semiparametric}{YoudenDRM}
}
