% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inference.R
\name{Youden}
\alias{Youden}
\title{Inference on the Youden index and optimal cutoff point}
\usage{
Youden(x, y, qt, r = NULL, totalSize, method = "optimal", CItype,
  level = 0.95)
}
\arguments{
\item{x}{observed sample from \eqn{F_0}.}

\item{y}{observed sample from \eqn{F_1}.}

\item{qt}{pre-sepecified basis functions of t in the exponential term. For example, \code{qt = c("t","log(t)")}.}

\item{r}{the value of the lower limit of detetion (LLOD). The default case is where no LLOD exists.}

\item{totalSize}{the total sample sizes \eqn{(n_0,n_1)}. When there is no LLOD (\code{r = NULL}), this argument is optional. See \code{help(DRM)} `Details'.}

\item{method}{the method used to fit the DRM. The options include \code{"glm"}, \code{"multinom"} and \code{"optimal"}.
The default setting is \code{"optimal"}. See \code{help(DRM)} for details.}

\item{CItype}{the method to be used for confidence interval construction. See `Details'.}

\item{level}{confidence level of the confidence interval. The default value is 0.95.}
}
\value{
The function returns a list containing the following components:
\itemize{
\item  Youden: the estimate, asymptotic standard deviation (ASD), and the confidence intervals (lower bound and uppper bound) of the Youden index.
When \code{CItype = "None"}, it only returns the estimate and ASD.
\item  cutoff: the estimate, ASD, and the confidence intervals (lower bound and uppper bound) of the optimal cutoff point.
When \code{CItype = "None"}, it only returns the estimate and ASD.
}
}
\description{
Estimate the Youden index and its corrreponding cutoff point under the density ratio model (DRM) as well as construct confidence intervals based on the provided data.
}
\details{
Let \eqn{F_0} and \eqn{F_1} denote the cumulative distribution functions of the healthy population and the diseased population, respectively.
The Youden index is defined as
\deqn{J = \max_x{F_0(x) - F_1(x)} = F_0(c) - F_1(c),}
where \eqn{c} is the corresponding cutoff point.

Under the DRM, the estimator \eqn{\hat c} of cutoff point is the solution to the equation
\deqn{\hat\alpha + \hat{\boldsymbol{\beta}}^\top \boldsymbol{q}(x) = 0,}
where \eqn{\hat\alpha} and \eqn{\hat{\boldsymbol{\beta}}} are the estimators of the paramters of DRM.
If multiple solution exist in the range of the observed samples, we choose the one that attains the maximum of \eqn{\hat F_0(\hat c) - \hat F_1(\hat c)} with
\eqn{\hat F_0} and \eqn{\hat F_1} being  the estimator of \eqn{F_0} under the DRM.
The explicit form of \eqn{\hat F_0} can be found in \insertCite{yuan2021semiparametric;textual}{YoudenDRM}.
If no solution exists  in the range of the observed samples, we find the \eqn{\hat c} by the definition.
With \eqn{\hat c}, the estimator of Youden index is \eqn{\hat J = \hat F_0(\hat c) - \hat F_1(\hat c)}.

The argument \code{CItype} refers to difference confidence intervals of the Younde index and cutoff point
under the DRMs in \insertCite{yuan2021semiparametric;textual}{YoudenDRM}.
\itemize{
\item  \code{"None"}: no confidence intervals for the Younde index and cutoff point is constructed;
\item  \code{"NA-DRM"}: the Wald-type confidence intervals for the Younde index and cutoff point based on the normal approximation;
\item  \code{"logit-DRM"}: the Wald-type confidence interval for the Youden index using logit transformation.
The Wald-type confidence interval for the cutoff point is contructed based on the normal approximation.
}
}
\examples{
#Example 1 (without LLOD)
#DMD is the dataset of data application in the paper.
x = DMD$CK[DMD$Status == 0]
y = DMD$CK[DMD$Status == 1]
Youden(x,y,qt = "t",CItype ="logit-DRM")
#Example 2 (with a LLOD)
#Date generation
set.seed(123456)
x = rlnorm(50, meanlog = 2.5, sdlog = sqrt(0.09))
y = rlnorm(50, meanlog = 2.87,sdlog = sqrt(0.25))
#Create the LLOD
r = qlnorm(0.15,meanlog = 2.5, sdlog = sqrt(0.09))
#Observed samples
x = x[x>=r]
y = y[y>=r]
# The basis function
qt = c("log(t)","log(t)^2")
Youden(x,y,qt,r,totalSize = c(50,50),CItype ="logit-DRM")
}
\references{
\insertRef{yuan2021semiparametric}{YoudenDRM}
}
