% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{goodnessFit}
\alias{goodnessFit}
\title{Goodness-of-fit test}
\usage{
goodnessFit(x, y, qt, r = NULL, totalSize, method = "optimal", B)
}
\arguments{
\item{x}{observed sample from \eqn{F_0}.}

\item{y}{observed sample from \eqn{F_1}.}

\item{qt}{pre-sepecified basis functions of t in the exponential term. For example, \code{qt = c("t","log(t)")}.}

\item{r}{the value of the lower limit of detetion (LLOD). The default case is where no LLOD exists.}

\item{totalSize}{the total sample sizes \eqn{(n_0,n_1)}. When there is no LLOD (\code{r = NULL}), this argument is optional. See \code{help(DRM)} `Details'.}

\item{method}{the method used to fit the DRM. The options include \code{"glm"}, \code{"multinom"} and \code{"optimal"}.
The default setting is \code{"optimal"}. See \code{help(DRM)} for details.}

\item{B}{number of boostrap reptitions.}
}
\value{
The function returns a list containing:
\itemize{
\item KSstat: the test statistic,
\item pval: p value of the test.
}
}
\description{
Goodness-of-fit test for the validity of the density ratio model (DRM) with a pre-specified basis function \code{qt}.
}
\details{
\insertCite{qin1997goodness;textual}{YoudenDRM} defined a  Kolmogorov-Smirnov-type test statistic as
\deqn{\Delta_n = \sup_{-\infty \leq x \leq \infty}   \sqrt{n}|\hat F_0(x) - \bar F_0(x)|,}
where \eqn{\hat F_0} refers to the estimator of \eqn{F_0} under the DRM
and \eqn{\bar F_0}  represents the empirical cumulative distribution function of \eqn{F_0}.
The explicit form of \eqn{\hat F_0} can be found in \insertCite{yuan2021semiparametric;textual}{YoudenDRM}.

We reject the null hypothesis that the DRM is satisfied if \eqn{\Delta_n} is greater than some critical value.
The limiting distribution of \eqn{\Delta_n} has a complicated form.
\insertCite{qin1997goodness;textual}{YoudenDRM} suggested to use a Bootstrap method to find the critical value.
}
\references{
\insertRef{qin1997goodness}{YoudenDRM}

\insertRef{yuan2021semiparametric}{YoudenDRM}
}
